rename database sub-namepaces (#1214)
