+++ python/lib/database_lib/visit_windows.py
@@ -6,7 +6,7 @@ from typing_extensions import deprecated
 __license__ = "GPLv3"
 
 
-@deprecated('Use `lib.db.model.visit_window.DbVisitWindow` instead')
+@deprecated('Use `lib.db.models.visit_window.DbVisitWindow` instead')
 class VisitWindows:
     """
     This class performs database queries for the VisitWindows table.
@@ -38,7 +38,7 @@ class VisitWindows:
         self.db = db
         self.verbose = verbose
 
-    @deprecated('Use `lib.db.query.visit.try_get_visit_window_with_visit_label` instead')
+    @deprecated('Use `lib.db.queries.visit.try_get_visit_window_with_visit_label` instead')
     def check_visit_label_exists(self, visit_label: str) -> bool:
         """
         Check if a visit label exists in the Visit_Windows database table.
