+++ python/tests/integration/scripts/test_run_dicom_archive_loader.py
@@ -1,7 +1,7 @@
 import subprocess
 
-from lib.db.query.config import set_config_with_setting_name
-from lib.db.query.mri_upload import get_mri_upload_with_patient_name
+from lib.db.queries.config import set_config_with_setting_name
+from lib.db.queries.mri_upload import get_mri_upload_with_patient_name
 from tests.util.database import get_integration_database_session
 from tests.util.file_system import check_file_tree
