+++ python/tests/unit/db/query/test_dicom_archive.py
@@ -4,10 +4,10 @@ import pytest
 from sqlalchemy import select
 from sqlalchemy.orm import Session as Database
 
-from lib.db.model.dicom_archive import DbDicomArchive
-from lib.db.model.dicom_archive_file import DbDicomArchiveFile
-from lib.db.model.dicom_archive_series import DbDicomArchiveSeries
-from lib.db.query.dicom_archive import (
+from lib.db.models.dicom_archive import DbDicomArchive
+from lib.db.models.dicom_archive_file import DbDicomArchiveFile
+from lib.db.models.dicom_archive_series import DbDicomArchiveSeries
+from lib.db.queries.dicom_archive import (
     delete_dicom_archive_file_series,
     get_dicom_archive_series_with_file_info,
     try_get_dicom_archive_with_study_uid,
